fix: 環境変数の設定をif文内で行う．cleanの冪等性を担保

+++ windows/clean.ps1
@@ -1,4 +1,8 @@
 # Scoopでインストールしたソフトウェアの削除とScoopのアンインストールを行う
-scoop uninstall scoop
-
-Remove-Item -Force -Recurse -Path "$HOEM\scoop\**"
+if (Get-Command scoop -ea SilentlyContinue) {
+  Write-Output "scoop をアンインストールします..."
+  scoop uninstall scoop
+  Remove-Item -Force -Recurse -Path "$HOEM\scoop\**"
+} else {
+  Write-Output "scoopがインストールされていません．スクリプトを終了します．"
+}
